Issue #45 : Designing a Settings Page
Created a settings page with the options to enable/disable notifications, change theme, adjust volume and enable/disable camera.
Linked the settings page to Profile #42
Created sidebars for channel navigation #43 #44
and linked them to each other, to the settings page and to the profile page.
To be implemented: adding text channels for direct and group messaging, and settings changes.