Issue #93 and #94 : Resolving CI errors
Changed assigned values in sidebar, group sidebar, login/signup and app files.
Updated functions for the sidebars for better backend integration.
Changed the layout so the group sidebar link can be replaced by the homepage.
This also ensures that the sidebars aren't displayed at the same time as the settings and profile.